Fixing the connectivity issue for the database containers
Created a seperate "CFConfig-actions.json" file for each of the application servers that would be used for the github actions and the "CFConfig.json" would be used for the TestUi app.
Modified the compose.yml file to use the "CFConfig-actions.json" file instead of "CFConfig.json".
In the "CFConfig-actions.json" files, changed the port for the database connectivity to the new port defined in the "compose.yml" file.
In the "CFConfig.json" files, changed the host to the name of the database service defined in the compose.yml file
Updated the "compose.yml" file to exposed different ports for the database services and used those ports in the "CFConfig-actions.json" file.
changed the line ending for the "init-db.sh" file from "CRLF" to "LF".