Going with the object oriented approach
Converted the application.cfc to not have includes and instead have all the necessary framework initialization functions inside it(Like onApplicationStart()).
Moved the inlcude for the "/app/config/app.cfm" to outside the onApplicationStart() function to make the csrf work on form submission.
Moved the Application.cfc and index.cfm back into the "public" folder.
Set the webroot to "Public" folder in all the server.json files.
Updated the event, model, plugins, views and controller paths in onapplicationstart.cfc.
Removed the functions.cfm file that was created to go back to the wheels 2.5 structure with includes as it is no longer required.