Updating the Introduction section of the guides and the docker's README.md file
Updated the docker's README.me file and removed the adobe 2016 server as it is no longer supported. Removed the VueJS mention from the file as we have rewrote the TestUI app in pure html, CSS and JS. Also added the missing ports for the available server versions.
Updated the URLs in the "routes.cfm" and "settings.cfm" file. Also added comment to explain the use of "set" function to set the datasource and etc.
Updated the wheels version in "onapplicationstart.cfc".
Fixed the broken links in the "congratulations.cfm" file.
In the files in the "guides/Introdution" folder, updated the files by updating the paths for the files, updating the code examples like adding 'application.wo' before the 'model("")' function, Fixing the broken links, Updating the available versions for lucee and adobe coldfusion for wheels 3.0.
In the "guides/introduction/upgrading.md", added the description of how to upgrade to the wheels 3.0