Updating the "handling requests with controllers" section of the guides
Updated the documents by fixing typos, updating directory structure, rewriting parts of the old documentation which had old conventions, updating broken links and adding "application.wo" to the necessary functions which are now in the "Global.cfc" file. Also, fixed some of the table design due to the change in the documents.
Also, removed "Andy Bellenie" as the core team member and added "Zain Ul Abideen"