Integrating methods from the wheels/model folder
Added initialization methods from the initialization.cfm to this file.
Added the $integrateComponents function which gets the cfc files from the provided path and passes those cfc file name to the $integrateFunctions function one by one by looping over the result.
Added the $integrateFunctions function which gets the functions from the provided cfc and then inserts that function into the current cfcs "variables" and "this" scope.
Add the call to the $integrateComponents in the constructor and passed it "wheels.model" to get cfc files from.