Converting cfm to cfcs and function injection
Converted the cfm for the controller and view components into cfcs.
Added the $integrateComponents function which gets the cfc files from the provided path and passes those cfc file name to the $integrateFunctions function one by one by looping over the result.
Added the $integrateFunctions function which gets the functions from the provided cfc and then inserts that function into the current cfcs "variables" and "this" scope.
Removed the unnecessary files